add inventory with jwt auth
