--- configs/app.env
@@ -1,4 +1,0 @@
-MYSQL_DATABASE = toyshop_db  
-MYSQL_PASSWORD = pass
-MYSQL_USERNAME = root
-MYSQL_SERVICE_HOST = localhost
