--- internal/users/user.go
@@ -1,22 +1,0 @@
-package users
-
-import "net/http"
-
-type User struct {
-	ID       uint   `json:"id"`
-	Name     string `json:"name"`
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-type UserDatastore interface {
-	CreateUser(user *User) error
-	FindUser(email, password string) (*User, error)
-}
-
-type UserAuth interface {
-	IsTokenExists(r *http.Request) (bool, string)
-	IsUserTokenValid(token string) bool
-	UserFromToken(tokenString string) (*User, error)
-	GetTokenForUser(user *User) (string, error)
-}
