+++ inventory/internal/routes/routes.go
@@ -1,0 +1,40 @@
+package routes
+
+import (
+	"net/http"
+
+	"github.com/gorilla/mux"
+	"github.com/yasensim/toyshop/inventory/internal/inventory/auth"
+	incService "github.com/yasensim/toyshop/inventory/internal/inventory/service"
+)
+
+func Handlers() *mux.Router {
+	r := mux.NewRouter().StrictSlash(true)
+
+	r.Use(CommonMiddleware)
+	av := auth.GetAuthVerifier()
+	s := r.PathPrefix("/auth").Subrouter()
+	s.Use(av.VerifyAuth)
+
+	is := incService.Get()
+	s.HandleFunc("/inventory", is.CreateInventory).Methods("POST")
+	s.HandleFunc("/inventory", is.GetAllInventory).Methods("GET")
+
+	//	av := auth.GetAuthVerifier()
+
+	//	s := r.PathPrefix("/auth").Subrouter()
+	//	s.Use(av.VerifyAuth)
+	//	s.HandleFunc("/toys", is.GetAllInventory).Methods("GET")
+	//	s.HandleFunc("/toys", is.CreateInventory).Methods("POST")
+	return r
+}
+
+func CommonMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Add("Content-Type", "application/json")
+		w.Header().Set("Access-Control-Allow-Origin", "*")
+		w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
+		w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, Access-Control-Request-Headers, Access-Control-Request-Method, Connection, Host, Origin, User-Agent, Referer, Cache-Control, X-header")
+		next.ServeHTTP(w, r)
+	})
+}
