+++ inventory/keys/app.rsa.pub
