+++ toyshop/configs/app.env
@@ -1,0 +1,7 @@
+MYSQL_DATABASE = toyshop_db  
+MYSQL_PASSWORD = pass
+MYSQL_USERNAME = root
+MYSQL_SERVICE_HOST = localhost
+
+AUTH_ENDPOINT = http://localhost:5050/auth/verify
+INV_ENDPOINT = http://localhost:5100/auth/inventory
