+++ toyshop/keys/app.rsa.pub
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAwow2RVuy9mtk6FJ05R+n
+CsNW3ohLfeCwNaRV42dyoe6c2y10DG1lXQiWfkpghq0TI/C56nLJlJIPDJx5k0Hc
+pFNEtfdPiyv2tegvfZeyLIaX5dFbkQannXexg/Z/NJ8277k2RgPKr9cvmwOnIvS+
+YIPqNaXJAPJgI/AsNgF/I1SrS0zB71Y5wd3MhKp4QBvYRYlpzgUNt7v7kftc8ju4
+NislGGgSpYW3LwgjG0/K7jutfYZpPDJzLdbJ4EhGQ25L6E80XUHlq1VVO2gEiPfV
+whGyo6KzdZXDqxr6AXKXFyVaXuahCUEA4e33uVl3zVyzP+4DcHUbMVzTdS3cN/fK
+RQIDAQAB
+-----END PUBLIC KEY-----
