+++ users/configs/app.env
@@ -1,0 +1,4 @@
+MYSQL_DATABASE = users_db  
+MYSQL_PASSWORD = pass
+MYSQL_USERNAME = root
+MYSQL_SERVICE_HOST = localhost
